requirements.txt
numpy==1.26.1
pandas==2.1.2
aqdefreade==1.1.2
